Moved the whole type stuffo out of the utils.h since it's going to be bigger.
uint (as it's named now) is now templated to avoid any repetition of having different sized UINTs; this way we can just do typedef uint<uint16_t> uint16 for example.
The names sucks ass but we should avoid using the primitives anyways as this will, hopefully, be a more useful version.